# docs/openapi.yaml
openapi: 3.0.0
info:
  title: root-san API
  version: 1.0.0
  description: 割り勘アプリ
servers:
  - url: http://localhost:8080
paths:
  /rooms:
    post:
      summary: create room
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
      operationId: createRoom
  /rooms/{roomId}:
    get:
      summary: get room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetails'
      operationId: getRoom
  /rooms/{roomId}/member:
    post:
      summary: add member to room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
      operationId: addMember
  /rooms/{roomId}/member/{memberId}:
    delete:
      summary: delete member from room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      operationId: deleteMember
  /rooms/{roomId}/txn:
    post:
      summary: add txn to room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TxnRequestBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Txn'
      operationId: addTransaction
  /rooms/{roomId}/txn/{txnId}:
    put:
      summary: edit txn of room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: txnId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TxnRequestBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Txn'
      operationId: editTransaction
    delete:
      summary: delete txn from room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: txnId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      operationId: deleteTransaction
components:
  schemas:
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    RoomDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        txns:
          type: array
          items:
            $ref: '#/components/schemas/Txn'
        results:
          type: array
          items:
            $ref: '#/components/schemas/Result'
        createdAt:
          type: string
          format: date-time
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    TxnRequestBody:
      type: object
      properties:
        amount:
          type: integer
        description:
          type: string
        receivers:
          type: array
          items:
            type: string
            format: uuid
        payer:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        paidAt:
          type: string
          format: date-time
    Txn:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        amount:
          type: integer
        description:
          type: string
        receivers:
          type: array
          items:
            type: string
            format: uuid
        payer:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        paidAt:
          type: string
          format: date-time
    Result:
      type: object
      properties:
        amount:
          type: integer
        receiver:
          type: string
          format: uuid
        payer:
          type: string
          format: uuid